+++ src/main/kotlin/net/pwall/json/JSONDeserializer.kt
@@ -708,6 +708,7 @@ object JSONDeserializer {
             val enclosingClass = enclosingType.classifierAsClass(this, pointer)
             val index = enclosingClass.typeParameters.indexOfFirst { it.name == typeParameter.name }
             return enclosingType.arguments.getOrNull(index)?.type ?:
+                    enclosingClass.typeParameters.getOrNull(index)?.upperBounds?.singleOrNull() ?:
                     fail("Can't create $simpleName - no type information for ${typeParameter.name}", pointer)
         }